Commit V24: Habilitar multidex y cambio en ProductosApp.dart

+++ lib/FbObjects/FbProducto.dart
@@ -3,7 +3,7 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 class FbProducto{
 
   final String nombre;
-  final double precio;
+  final int precio;
   final String imagen;
 
   FbProducto({
